pyproject.toml
[tool.poetry]
name = "prelevement-deau-airflow"
version = "0.1.0"
description = ""
authors = ["anislerouge <anislerouge@gmail.com>"]
readme = "README.md"
packages = [{include = "dags"}, {include = "utils"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-dotenv = "^1.0.1"
pre-commit = "^3.7.0"
init = "^0.1.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
apache-airflow = "^2.9.0"
psycopg2-binary = "^2.9.9"
boto3 = "^1.34.106"
requests = "^2.32.2"
pandas = "^2.2.2"
odfpy = "^1.4.1"
xlrd = "^2.0.1"
openpyxl = "^3.1.3"
pytest = "^8.2.2"
python-dateutil = "^2.9.0.post0"
alembic = "^1.13.2"
geoalchemy2 = "^0.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
